requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
altair>=5.0.0
groq>=0.4.0
python-dotenv>=1.0.0
streamlit-extras>=0.3.0
